// src/app/core/models/navigationtab.ts
export interface LogoCinema {
    maHeThongRap: string;
    tenHeThongRap: string;
    biDanh: string;
    logo: string;
}

// export interface SystemCinema {
//     maCumRap: string;
//     tenCumRap: string;
//     diaChi: string;
//     danhSachRap: DanhSachRap[];
// }
// export interface DanhSachRap {
//     maRap: number;
//     tenRap: string;
// }


export interface SystemCinema {
    lstCumRap: LstCumRap[];
    maHeThongRap: string;
    tenHeThongRap: string;
    logo: string;
    mahom: string;
}
export interface LstCumRap {
    danhSachPhim: DanhSachPhim[];
    maCumRap: string;
    tenCumRap: string;
    diaChi: string;
}
export interface DanhSachPhim {
    lstLichChieuTheoPhim: LstLichChieuTheoPhim[];
    maPhim: number;
    tenPhim: string;
    hinhAnh: string;
}
export interface LstLichChieuTheoPhim {
    maLichChieu: number;
    maRap: string;
    tenRap: TenRap;
    ngayChieuGioChieu: Date;
    giaVe: number;
}
export enum TenRap {
    Rạp1 = "Rạp 1",
    Rạp10 = "Rạp 10",
    Rạp2 = "Rạp 2",
    Rạp3 = "Rạp 3",
    Rạp4 = "Rạp 4",
    Rạp5 = "Rạp 5",
    Rạp6 = "Rạp 6",
    Rạp7 = "Rạp 7",
    Rạp8 = "Rạp 8",
    Rạp9 = "Rạp 9",
}
export interface MovieSchedule {
    giaVe: number;
    maLichChieu: number;
    maRap: string;
    ngayChieuGioChieu: string;
    tenRap: string;
}